requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
python-dotenv
psycopg2-binary
twilio
